+++ CHANGELOG.md
@@ -123,3 +123,4 @@
 - 2023-10-07: Pagination gp, session, deals, scheduler, +3
 - 2023-10-07: Impl session, scheduler, deals, qualifier, +5 query
 - 2023-10-08: Metrics gp, qualifier, scheduler, session, +5
+- 2023-10-09: Pagination qualifier, session, deals, gp, +4